+++ CHANGELOG.md
@@ -31,6 +31,14 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Internal
   - replace all texconv usage to reduce number of tools
 
+## [0.5.1]
+
+### Changed
+
+- Improved outline name detection for .cf7 feedback definitions
+- Support `${annoRda}` in command line mode
+- Fixed issue with some features not loading properly
+
 ## [0.5.0]
 
 ### Added